Add router, privacy policy, some more info and links in about dialog
