+++ frontend/src/router.ts
@@ -1,14 +1,23 @@
 import { createRouter, createWebHistory } from 'vue-router'
-//import Main from '@/views/Main.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
-    // {
-    //   path: '/',
-    //   name: 'main',
-    //   component: Main
-    // }
+    {
+      path: '/',
+      name: 'main',
+      component: () => import("@/views/Main.vue")
+    },
+    {
+      path: '/privacy',
+      name: 'privacy',
+      component: () => import("@/views/Privacy.vue")
+    },
+    {
+      path: '/:pathMatch(.*)*',
+      name: 'main',
+      component: () => import("@/views/Main.vue")
+    }
   ]
 })
